refactor(barcode-scanner): move useBarcodeScanner hook to features/receiving and update imports
