+++ src/app/(dashboard)/receiving/page.tsx
@@ -31,7 +31,7 @@ import {
 import { usePagination, useNotifications } from "@/contexts/app-context";
 import { CustomTable, TableColumn } from "@/components/common/custom-table";
 import { BarcodeScannerDialog } from "@/components/common/barcode-scanner-dialog";
-import { useBarcodeScanner } from "@/hooks/use-barcode-scanner";
+import { useBarcodeScanner } from "@/features/receiving/hooks/use-barcode-scanner";
 import { usePdfReport } from "@/hooks/use-pdf-report";
 import { ProtectedRoute } from "@/components/common/protected-route";
