+++ src/features/receiving/hooks/use-barcode-scanner.tsx
